Update Client to contain an http.Client
This http.Client will be used to make testing possible.
Presently, the important doMethod and doMethodAndParseBody use the
http.DefaultClient, with no way of overriding this.
Now that Client relies on using an http.Client, these methods can also
use that client. So mocking out the http.Client is now available.